feat(core): validation factory is not longuer supported in Micronaut 4
