+++ core/src/main/java/io/kestra/core/validations/CronExpression.java
@@ -1,10 +1,11 @@
 package io.kestra.core.validations;
 
+import io.kestra.core.validations.validator.CronExpressionValidator;
 import javax.validation.Constraint;
 import java.lang.annotation.*;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = { })
+@Constraint(validatedBy = CronExpressionValidator.class)
 public @interface CronExpression {
     String message() default "invalid cron expression ({validatedValue})";
 }
