+++ core/src/main/java/io/kestra/core/validations/DateFormat.java
@@ -1,11 +1,12 @@
 package io.kestra.core.validations;
 
+import io.kestra.core.validations.validator.DateFormatValidator;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import javax.validation.Constraint;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = { })
+@Constraint(validatedBy = DateFormatValidator.class)
 public @interface DateFormat {
     String message() default "invalid date format value";
 }
