+++ core/src/main/java/io/kestra/core/validations/JsonString.java
@@ -1,11 +1,12 @@
 package io.kestra.core.validations;
 
+import io.kestra.core.validations.validator.JsonStringValidator;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import javax.validation.Constraint;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = { })
+@Constraint(validatedBy = JsonStringValidator.class)
 public @interface JsonString {
     String message() default "invalid json ({validatedValue})";
 }
