+++ core/src/main/java/io/kestra/core/validations/Schedule.java
@@ -1,11 +1,12 @@
 package io.kestra.core.validations;
 
+import io.kestra.core.validations.validator.ScheduleValidator;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import javax.validation.Constraint;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = { })
+@Constraint(validatedBy = ScheduleValidator.class)
 public @interface Schedule {
     String message() default "invalid schedule ({validatedValue})";
 }
