+++ core/src/main/java/io/kestra/core/validations/SwitchTaskValidation.java
@@ -1,11 +1,12 @@
 package io.kestra.core.validations;
 
-import javax.validation.Constraint;
+import io.kestra.core.validations.validator.SwitchTaskValidator;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
+import javax.validation.Constraint;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = { })
+@Constraint(validatedBy = SwitchTaskValidator.class)
 public @interface SwitchTaskValidation {
     String message() default "invalid Switch task";
 }
