+++ core/src/main/java/io/kestra/core/validations/WorkingDirectoryTaskValidation.java
@@ -1,11 +1,12 @@
 package io.kestra.core.validations;
 
+import io.kestra.core.validations.validator.WorkingDirectoryTaskValidator;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import javax.validation.Constraint;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = { })
+@Constraint(validatedBy = WorkingDirectoryTaskValidator.class)
 public @interface WorkingDirectoryTaskValidation {
     String message() default "invalid WorkingDirectory task";
 }
